fixed big bug.....
so it all started when I noticed that if a target along the way (far
into the chain) failed than you would have to start over from scratch.
Parallel builds had support to make incremental updated to the shastore,
but the serial build did not. So I added that.
SECOND, the parallel build's ability to do that was compromised because
I forgot to update the in_mem_shas correctly (using dictionary with
sha key). So I fixed that.
THIRD, for some strange reason it only updated the shas if they were
in the from_store. I don't think it needs to. Why did I do that?
FOURTH, it also had to incrementally shastore update the dependencies
(not just the output) so I fixed that.
FIFTH, there was a bug in one on the functional tests; in "sake force
'compile objects', it depended on a file that wasn't available so it
should have failed. However, at the time, we weren't checking the shas
of any of the DEPENDENCIES, so the error slipped through, so I fixed
that.